Move `ledger-api-transactions-buffer-max-size` to Config from ParticipantConfig
changelog_begin
changelog_end

+++ ledger/participant-integration-api/src/main/scala/platform/store/cache/EventsBuffer.scala
@@ -34,7 +34,7 @@ import scala.math.Ordering
   * @tparam ENTRY The entry buffer type.
   */
 final class EventsBuffer[ENTRY](
-    maxBufferSize: Long,
+    maxBufferSize: Int,
     metrics: Metrics,
     bufferQualifier: String,
     isRangeEndMarker: ENTRY => Boolean,
